Add location dataset and create GPS map image
When doing this, I kept getting a weird error message that R encountered a fatal error and needs to restart. This occurred only with this file in the project. After this error continued to appear, I uninstalled R and RStudio and then reinstalled the newest version of them. Everything seems to be working now at least.